shift hover to app drawer
