+++ packages/common/src/ui/components/navigation/AppNavLink/AppNavLink.test.tsx
@@ -89,6 +89,7 @@ describe('AppNavLink', () => {
                   icon={<TruckIcon />}
                   text="Distribution"
                   end={false}
+                  inactive
                 />
               </Box>
             }
